wip: uvsphere mesh gen and pbr shader first pass

+++ src/maths/primitives.h
@@ -9,5 +9,5 @@
 geometry_data geo_create_plane(f32x2 extents);
 geometry_data geo_create_cuboid(f32x3 extents);
 geometry_data geo_create_cylinder(f32 radius, f32 height, u32 resolution);
-geometry_data geo_create_uvsphere(f32 radius, f32 north_south_lines, f32 east_west_lines);
+geometry_data geo_create_uvsphere(f32 radius, u32 north_south_lines, u32 east_west_lines);
 geometry_data geo_create_icosphere(f32 radius, f32 n_subdivisions);
